added case service and case.ts
