+++ src/app/case/case.ts
@@ -1,0 +1,7 @@
+export interface Case {
+  name: String 
+  student_id: String
+  cell_number: String
+  quarantine_location: String
+  quarantine_end_date: String
+}
